overlay|refactor: Regroup overlay's flat src modules by role
Moves every module out of `overlay`'s flat `src/` root into a directory named for its role: `types.ts` into `modes/`, which owns the vocabulary every mode returns; `formatReport.ts` and `formatJsonError.ts` into a new `reporting/`; and `utils/pluralize.ts` into `portable/`, alongside the generic code readyup and compositor keep there. `index.ts` and `overlay.ts` stay at the root, naming the package rather than a role within it.

`pluralize.ts` picks up the doc descriptions its readyup twin already carries.
