+++ packages/overlay/src/modes/create.ts
@@ -4,8 +4,8 @@ import { parseStatus } from '../chezmoi/parseStatus.ts';
 import { readStatus } from '../chezmoi/readStatus.ts';
 import type { ChezmoiContext } from '../chezmoi/runChezmoi.ts';
 import { runChezmoiStreamed } from '../chezmoi/runChezmoi.ts';
-import type { OverlayResult } from '../types.ts';
 import { countOutcome, partitionStatus } from './buildEntries.ts';
+import type { OverlayResult } from './types.ts';
 
 /**
  * Non-clobbering convergence: create missing entries (`A`), perform native deletions (`D`), run `run_` scripts (`R`),
