+++ packages/overlay/src/reporting/formatJsonError.ts
